Apply codestyle
