+++ src/main/kotlin/com/featurevisor/sdk/Emitter.kt
@@ -3,6 +3,7 @@ package com.featurevisor.sdk
 import com.featurevisor.types.EventName
 
 class Emitter {
+
     private val listeners = mutableMapOf<EventName, (Array<out Any>) -> Unit>()
 
     fun addListener(event: EventName, listener: (Array<out Any>) -> Unit) {
