+++ src/main/kotlin/com/featurevisor/sdk/Instance+Activation.kt
@@ -2,7 +2,7 @@ package com.featurevisor.sdk
 
 import com.featurevisor.types.AttributeValue
 import com.featurevisor.types.Context
-import com.featurevisor.types.EventName
+import com.featurevisor.types.EventName.ACTIVATION
 import com.featurevisor.types.FeatureKey
 import com.featurevisor.types.VariationValue
 
@@ -15,13 +15,13 @@ fun FeaturevisorInstance.activate(featureKey: FeatureKey, context: Context = emp
         .filter { it.capture == true }
 
     attributesForCapturing.forEach { attribute ->
-        if (finalContext[attribute.key] != null) {
-            captureContext[attribute.key] = context[attribute.key]!!
+        finalContext[attribute.key]?.let {
+            captureContext[attribute.key] = it
         }
     }
 
     emitter.emit(
-        EventName.ACTIVATION,
+        ACTIVATION,
         featureKey,
         variationValue,
         finalContext,
