+++ src/main/kotlin/com/featurevisor/sdk/Instance+Status.kt
@@ -2,4 +2,6 @@ package com.featurevisor.sdk
 
 data class Statuses(var ready: Boolean, var refreshInProgress: Boolean)
 
-fun FeaturevisorInstance.isReady(): Boolean = statuses.ready
+internal fun FeaturevisorInstance.isReady(): Boolean {
+    return statuses.ready
+}
