+++ src/main/kotlin/com/featurevisor/sdk/Instance+Variation.kt
@@ -4,7 +4,7 @@ import com.featurevisor.types.Context
 import com.featurevisor.types.FeatureKey
 import com.featurevisor.types.VariationValue
 
-fun FeaturevisorInstance.getVariation(featureKey: FeatureKey, context: Context): VariationValue? {
+internal fun FeaturevisorInstance.getVariation(featureKey: FeatureKey, context: Context): VariationValue? {
     val evaluation = evaluateVariation(featureKey, context)
     return when {
         evaluation.variationValue != null -> evaluation.variationValue
