--- src/main/kotlin/com/featurevisor/types/EventName.kt
@@ -1,8 +1,0 @@
-package com.featurevisor.types
-
-enum class EventName {
-    READY,
-    REFRESH,
-    UPDATE,
-    ACTIVATION,
-}
